requirements.txt
# ── sitrep_tracker ────────────────────────────────────────────────────────────
click>=8.0
rich>=13.0

# ── scripts/build_pdfs.py ─────────────────────────────────────────────────────
markdown>=3.4
websocket-client>=1.6

# ── scripts/build_*_slide.py (PowerPoint generation) ─────────────────────────
python-pptx>=0.6.21

# ── apps (readiness_tracker, exam_analytics, aar_aggregator) ─────────────────
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
pydantic>=2.7
python-multipart>=0.0.9
streamlit>=1.35
plotly>=5.22
pandas>=2.2

# ── dev / testing ─────────────────────────────────────────────────────────────
pytest>=7.0
pytest-mock>=3.0
httpx>=0.27
